requirements.txt
Flask==0.11.1
Flask-SQLAlchemy==2.1
Flask-Testing==0.5.0
GitHub-Flask==3.1.3
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysql-connector-python-rf==2.2.2
pytest==3.0.1
pytest-cov==2.3.1
pytest-flask==0.10.0
requests==2.11.1
six==1.10.0
SQLAlchemy==1.0.14
Werkzeug==0.11.10
pymysql==0.7.11
WTForms==2.1
passlib==1.7.1
pyyaml==3.12
parsel==1.1
